move findMapStyleUrl
