+++ src/components/views/location/LocationShareErrors.ts
@@ -1,0 +1,5 @@
+export enum LocationShareError {
+    MapStyleUrlNotConfigured = 'MapStyleUrlNotConfigured',
+    MapStyleUrlNotReachable = 'MapStyleUrlNotReachable',
+    Default = 'Default'
+}
